Made Main, Counter, and IncrementForm into routes and pulled the logic into the components that used them.
