+++ src/App.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
-import Main from './components/Main/Main';
+import Routes from './routes';
 
 
 class App extends Component {
   render() {
     return (
       <div className="App">
-        <Main />
+        <Routes />
       </div>
     );
   }
